fix: Dangling Tesseracts in e2e tests (#51)
#### Relevant issue or PR
Fix #34 

#### Description of changes
>I went with adding try-finally when meaningful for the following
reasons:
>
> * The tests right now are mostry using the cli with various options;
switching to the python sdk there would make some stuff untestable, and
further reduce coverage to the cli (python sdk is simpler and more
unit-tested).
> * One could write a parametrized fixture which is basically a function
that pipes some args to cli_runner.invoke( , yields, and then does the
teardown of that stuff -- but each call to invoke is slightly-different
enough that the fixture itself would become a bit too large for my
taste, and imho there is value to seeing the actual command which ran
spelled out explicitly, at least in tests.
>
>Minor observation: tbh some of the stuff here would be fine as a unit
test rather than an e2e one (test_tesseract_serve_ports_error, for
example), but I'm fine with leaving it be for the time being.

https://github.com/pasteurlabs/tesseract-core/issues/34#issuecomment-2687556316

#### Testing done
CI

#### License

- [ x] By submitting this pull request, I confirm that my contribution
is made under the terms of the [Apache 2.0
license](https://pasteurlabs.github.io/tesseract/LICENSE).
- [ x] I sign the Developer Certificate of Origin below by adding my
name and email address to the `Signed-off-by` line.

<details>
<summary><b>Developer Certificate of Origin</b></summary>

```text
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.


Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
```

</details>

Signed-off-by: Dr. Alessandro Angioi
<alessandro.angioi@simulation.science>

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>