feat: Remove docker_py usage in favor of custom client that uses Docker CLI (#33)
#### Relevant issue or PR
Fixes #10 
Enables #8

#### Description of changes
- Created a new `docker_client.py` file that contains a
`CLIDockerClient` class that is a custom python wrapper over
`docker_cli` that mimics the interface of `docker_py`.
- The file has 3 additional classes, `Container`, `Compose`, and
`Image`.
- Removed a lot of functions from existing files (all calls to
subprocess) and docker errors. Migrated them to the `docker_client.py`
and replaced errors with our new errors.
- `teardown` function in cli has the most amount of changes as handling
of `--all` flag is moved to `engine.teardown` instead as we no longer
need to filter through all the projects in docker and parse which ones
belong to Tesseract as our `DockerWrapper` class already knows all the
projects that are Tesseracts.
- Renamed `_display_container_meta` to `_display_project_meta` as I
think it's more fitting to the function since we are listing the
metadata of all the containers in a project, aka the project's metadata.
- Modified some unit tests to fit the new mocked class.

BREAKING CHANGE: Tesseract Core does not clear build caches anymore.
That means, the option `--keep-build-cache` and environment variable
`TESSERACT_KEEP_BUILD_CACHE` are now removed, and users are advised to
clean up Docker resources manually as needed (e.g. via `docker system
prune`).

#### Testing done
Unit tests + Manually tested commands:
`serve` `build` `ps` `list` `teardown` `teardown --all` `apidoc`

#### License

- [x] By submitting this pull request, I confirm that my contribution is
made under the terms of the [Apache 2.0
license](https://pasteurlabs.github.io/tesseract/LICENSE).
- [x] I sign the Developer Certificate of Origin below by adding my name
and email address to the `Signed-off-by` line.

<details>
<summary><b>Developer Certificate of Origin</b></summary>

```text
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.


Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
```

</details>

Signed-off-by: [Angela Ko] <angela.ko@simulation.science>

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>