doc: Tesseract Example Gallery (#149)
#### Relevant issue or PR


#### Description of changes
Added arm runner and example directory for arm64, along with updated
website gallery for examples.

Use sphinx title cards to showcase examples
<img width="1138" alt="image"
src="https://github.com/user-attachments/assets/8aafde0c-825a-48ee-b136-19a56d22b066"
/>

Due to the nature of Tesseracts and it being not purely python code, it
is difficult to use features such as sphinx testing and sphinx gallery,
as these are primarily for python changes. Even in the cases where you
could possibly add in non-python code for display, it is difficult
because the auto-generating nature of sphinx gallery means we would need
[heavy handhelding to tell sphinx which files to generate and which to
skip](https://sphinx-gallery.github.io/stable/configuration.html#build-pattern).
Furthermore, testing requires environment modification such as running
and building the tesseract on certain systems, etc.

Sphinx-gallery and testing also have the prereq that your logic is
executable in one file, which is also not the case for Tesseracts.

Because of that, I ended up going with the original decision of just
using our test-examples logic and having the logic in the how-to guides
live in `examples` for testing, but modified so that it would look more
visually like `sphinx-gallery`.

#### Testing done
Unit Testing

#### License

- [ x] By submitting this pull request, I confirm that my contribution
is made under the terms of the [Apache 2.0
license](https://pasteurlabs.github.io/tesseract/LICENSE).
- [ x] I sign the Developer Certificate of Origin below by adding my
name and email address to the `Signed-off-by` line.

<details>
<summary><b>Developer Certificate of Origin</b></summary>

```text
Developer Certificate of Origin
Version 1.1

Copyright (C) 2004, 2006 The Linux Foundation and its contributors.

Everyone is permitted to copy and distribute verbatim copies of this
license document, but changing it is not allowed.


Developer's Certificate of Origin 1.1

By making a contribution to this project, I certify that:

(a) The contribution was created in whole or in part by me and I
    have the right to submit it under the open source license
    indicated in the file; or

(b) The contribution is based upon previous work that, to the best
    of my knowledge, is covered under an appropriate open source
    license and I have the right under that license to submit that
    work with modifications, whether created in whole or in part
    by me, under the same open source license (unless I am
    permitted to submit under a different license), as indicated
    in the file; or

(c) The contribution was provided directly to me by some other
    person who certified (a), (b) or (c) and I have not modified
    it.

(d) I understand and agree that this project and the contribution
    are public and that a record of the contribution (including all
    personal information I submit with it, including my sign-off) is
    maintained indefinitely and may be redistributed consistent with
    this project or the open source license(s) involved.
```

</details>

Signed-off-by: Angela Ko  <angela.ko@simulation.science>

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>