feat(sdk): expose no compose in Python API (#223)
This is necessary to allow using `.serve()` and `.from_image()` when
`docker-compose`/`podman-compose` are missing.

### Changes made
`no_compose` argument exposed in `Tesseract.from_image()` and
`Tesseract.serve()`. If `True` then this is passed to `engine.serve()`
which already provides the translation of this into the CLI with the
`--no-compose` option.

#### Testing done
I did not add a test as I do not understand if this will be meaningful
with `pytest-mock`. I will test on NERSC next week, @josiahbjorgaard let
me know if this works on your side!

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>