fix: add volume statements to dockerfile to ensure logs / metrics / data are always written to a volume (#270)
#### Relevant issue or PR
n/a

#### Description of changes
Without this, logs / metrics / artifacts are written to the container
filesystem if `--output-path` is not passed. This causes data to
accumulate over time with no straightforward way to read it. By adding
`VOLUME` to the Dockerfile we make sure there is always a Docker volume
to write to, even if not specifically requested.

#### Testing done
none