fix: Use `--input-path`/`--output-path` as `base_dir` in `json+binref` encoder (#304)
#### Relevant issue or PR
When passing `json+binref` encoded payloads to a tesseract, the
`base_dir` (which stores the .bin files) was not set.
This would not be an issue if the payload would be encoded with a proper
`InputFileReference`, but I don't think we can assume this in all cases.

#### Description of changes
Use input path/output path as the default for `base_dir` (if it is not
explicitly set)

#### Testing done
CI passes.

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>