fix(sdk): Cast sets to tuples in Pytorch template when used as positional args (#347)
The iteration order of a `set` is not guaranteed but is relied upon by
the filtered positional arguments. To ensure behaviour is as expected
these are now cast to `tuple`s. This already happens in the jax template
to satisfy jit.

#### Relevant issue or PR
- #344 

#### Description of changes
Cast `jvp_inputs` `jac_inputs` `vjp_inputs` to `tuple`s . Output paths
are not cast as these are used as keys to a dictionary so order is not
important.

#### Testing done
CI passes, I should maybe add some specific tests to catch edge cases.
Let me know if you'd like this!