refactor!: Use uv to install python virtual env (#353)
#### Description of changes
Uses `uv` rather than `pip` to install the virtual environment in the
docker image. This should be transparent to users: the virtual
environment created at the end is the same.

On a test tesseract build, I observed the installation of dependencies
into the virtual env speed up by ~6x. There's another case I've run into
where pip is actually unable to resolve the dependencies and gets stuck;
I've been using this branch to get around that.

I referred to the following documentation page when making this change:
https://docs.astral.sh/uv/guides/integration/docker/

#### Testing done
- Added an example tesseract containing a custom `pip install` build
step and corresponding test case
- Manual testing: I've been using this patch for my own tesseract builds
for about a month, and seen no issues.

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>