feat: add example for differentiable quadratic programming (QP) solver (#345)
#### Relevant issue or PR
n/a

#### Description of changes
This adds an example which makes use of the differentiable QP solver
[qpax](https://github.com/kevin-tracy/qpax). In the example, a
constraint in a simplistic 2d QP is adjusted so that the solution of the
QP matches a target point. The constraint is updated using the gradients
of the QP solution provided by `qpax`.

#### Testing done
Added a test for `apply` endpoint in `test_examples.py`. The script to
generate the animation runs locally.

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>
Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>