fix: catch identical mounted volumes (#337)
#### Relevant issue or PR
This prevents `parsed_volume` entries being overwritten when the same
source folder is given to multiple volumes, or `input-path` and
`output-path`. Instead, a descriptive `ValueError` is raised.
Previously, this would not mount the duplicate volume, which results in
an opaque `FileNotFoundError` later on.

Instead of this behavior, we might want to enable the mounting of the
same folder to multiple locations inside the tesseract. This would
potentially entail some more checks of file names, permissions, etc to
prevent unexpected behavior.

#### Description of changes
`engine._parse_option` is modified to raise an error is duplicate
`source` folders are encountered. Also, in `engine.run_tesseract` and
`engine.serve`, it is checked whether `input_path` or `output_path` are
already denoted as parsed volumes.

Another, vaguely related fix: When using the CLI, `output-path` is
created if it does not already exists. This is not the case when using
`from_image` or `from_api`. Changes to `sdk/tesseract.py` fix this. Now,
running `examples/filereference/test_tesseract.py` does no longer fail
without first creating the `output` directory. (I can remove this in
case there is some reason for why the path is not created for the
`from_*` interfaces.)

#### Testing done
Added checks to existing tests for `engine.run_tesseract` and
`engine.serve`, to make sure error is raised when identical folders are
in volumes or in input path and output path.

---------

Co-authored-by: Alessandro Angioi <alessandro.angioi@simulation.science>