fix: ensure log messages aren't lost when application exits (#392)
#### Relevant issue or PR
#356, #357

#### Description of changes
This fixes the observed issue of missing logs (printed messages +
tracebacks) in some scenarios, specifically when the interpreter crashes
/ exits before corresponding I/O pipes are exhausted.

Also renames `LogPipe` -> `TeePipe` because that's really what it does
(teeing).

The primary difference is that `TeePipe` now blocks the main thread when
exiting while there are messages waiting in the pipe. Only when nothing
has been read for X seconds do we continue with shutdown. This ensures
that all queued up messages are forwarded properly.

#### Testing done
new test + manual check that messages are now shown as expected on the
reproducer from #356.