chore: 🚢 release v1.2.0 (#398)
This PR contains the generated changelog for the release v1.2.0.

⚠️ **Merging this PR will immediately trigger a new release**. ⚠️

To specify additional release notes, please edit this comment after the
following line.

---

This release brings several quality-of-life improvements to Tesseract
Core and many new examples and showcases in our
[docs](https://docs.pasteurlabs.ai/projects/tesseract-core/latest/).

**Breaking change:** Support for Python 3.9 is now removed (has been EOL
since October 2025). If you rely on Python 3.9 please pin to
`tesseract_core<1.2.0` and upgrade as soon as possible.

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>