fix: reachability check uses mlflow username/password if provided (#416)
<!--
Please use a PR title that conforms to *conventional commits*:
"<commit_type>: Describe your change"; for example: "fix: prevent race
condition". Some other commit types are: fix, feat, ci, doc, refactor...
For a full list of commit types visit
https://www.conventionalcommits.org/en/v1.0.0/
-->

#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. Only link something if it is directly relevant. -->

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->

If auth is present in MLFlow, we need to pass the creds in via
environment variables
(https://mlflow.org/docs/latest/self-hosting/security/basic-http-auth/#using-environment-variables).
While this works for when we use the mlflow package, the reachability
check just uses `requests`. This adds the username/password to the
requests.

Note: I thought about using the `/health` endpoint of mlflow instead
(which is un-authed on mlflow) but it is (a) undocumented so might
change and (b) doesn't replicate actual usage as well.
We also don't use the traditional tesseract config package (prepand
config with `TESSERACT`) for the same reason (want to replicate the
actual process the mlflow package will use when it makes requests).

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

Local testing