feat(sdk): change default Python SDK `output_format` to b64 for performance reasons (#422)
#### Relevant issue or PR
N/A but related to #423

#### Description of changes
Change default `output_format` to json+base64 for performance reasons results in 10x serialisation speedup for 1E6 array size with vector add.

Officially drop support for json+binref (mainly because implementation not entirely trivial and performance benefits not clear) this may change in the future.

#### Testing done

- [x] CI passes
- [x] Updated assert_called_with
- [x] Add dtype tests
- [x] Parametrize test_tesseract_sdk over output_format
- [x] Performance for array sizes of 1E^ improved by >10x for 

---------

Co-authored-by: Jonathan Brodrick <jonathanbrodrick@gmail.com>
Co-authored-by: Dion Häfner <dion.haefner@simulation.science>