feat: add ability to specify mlflow tag for tesseract (#426)
<!--
Please use a PR title that conforms to *conventional commits*:
"<commit_type>: Describe your change"; for example: "fix: prevent race
condition". Some other commit types are: fix, feat, ci, doc, refactor...
For a full list of commit types visit
https://www.conventionalcommits.org/en/v1.0.0/
-->

#### Relevant issue or PR
<!-- If the changes resolve an issue or follow some other PR, link to
them here. Only link something if it is directly relevant. -->

#### Description of changes
<!-- Add a high-level description of changes, focusing on the *what* and
*why*. -->

This adds the ability to specify mlflow tags for Tesseract runs via the
`TESSERACT_MLFLOW_TAGS` env variable. This is useful to embed additional
metadata into the MLFlow run.

n.b. As we want to ensure that all runs have the tags if set, we need to
make a small change here to check if there is an active run. Previously,
if there was no an active run and you logged something, it would
implicitly create one. That is a problem as it will not include the
tags. Instead we now explicitly check if there an active run; and, if
not, we create one with the tags.

#### Testing done
<!-- Describe how the changes were tested; e.g., "CI passes", "Tested
manually in stagingrepo#123", screenshots of a terminal session that
verify the changes, or any other evidence of testing the changes. -->

Local and unit testing

---------

Co-authored-by: Dion Häfner <dion.haefner@simulation.science>